Crate rusoto_fms

AWS Firewall Manager
This is the AWS Firewall Manager API Reference. This guide is for developers who need detailed information about the AWS Firewall Manager API actions, data types, and errors. For detailed information about AWS Firewall Manager features, see the AWS Firewall Manager Developer Guide.

If you're using the service, you're probably looking for FmsClient and Fms.

Structs

AssociateAdminAccountRequest

ComplianceViolator
Details of the resource that is not protected by the policy.

DeleteNotificationChannelRequest

DeletePolicyRequest

DisassociateAdminAccountRequest

EvaluationResult
Describes the compliance status for the account. An account is considered non-compliant if it includes resources that are not protected by the specified policy.

FmsClient
A client for the FMS API.

GetAdminAccountRequest

GetAdminAccountResponse

GetComplianceDetailRequest

GetComplianceDetailResponse

GetNotificationChannelRequest

GetNotificationChannelResponse

GetPolicyRequest

GetPolicyResponse

GetProtectionStatusRequest

GetProtectionStatusResponse

ListComplianceStatusRequest

ListComplianceStatusResponse

ListMemberAccountsRequest

ListMemberAccountsResponse

ListPoliciesRequest

ListPoliciesResponse

Policy
An AWS Firewall Manager policy.

PolicyComplianceDetail
Describes the non-compliant resources in a member account for a specific AWS Firewall Manager policy. A maximum of 100 entries are displayed. If more than 100 resources are non-compliant, EvaluationLimitExceeded is set to True.

PolicyComplianceStatus
Indicates whether the account is compliant with the specified policy. An account is considered non-compliant if it includes resources that are not protected by the policy.

PolicySummary
Details of the AWS Firewall Manager policy.

PutNotificationChannelRequest

PutPolicyRequest

PutPolicyResponse

ResourceTag
The resource tags that AWS Firewall Manager uses to determine if a particular resource should be included or excluded from protection by the AWS Firewall Manager policy. Tags enable you to categorize your AWS resources in different ways, for example, by purpose, owner, or environment. Each tag consists of a key and an optional value, both of which you define. Tags are combined with an "OR." That is, if you add more than one tag, if any of the tags matches, the resource is considered a match for the include or exclude. Working with Tag Editor.

SecurityServicePolicyData
Details about the security service that is being used to protect the resources.

Enums

AssociateAdminAccountError
Errors returned by AssociateAdminAccount

DeleteNotificationChannelError
Errors returned by DeleteNotificationChannel

DeletePolicyError
Errors returned by DeletePolicy

DisassociateAdminAccountError
Errors returned by DisassociateAdminAccount

GetAdminAccountError
Errors returned by GetAdminAccount

GetComplianceDetailError
Errors returned by GetComplianceDetail

GetNotificationChannelError
Errors returned by GetNotificationChannel

GetPolicyError
Errors returned by GetPolicy

GetProtectionStatusError
Errors returned by GetProtectionStatus

ListComplianceStatusError
Errors returned by ListComplianceStatus

ListMemberAccountsError
Errors returned by ListMemberAccounts

ListPoliciesError
Errors returned by ListPolicies

PutNotificationChannelError
Errors returned by PutNotificationChannel

PutPolicyError
Errors returned by PutPolicy

Traits

Fms
Trait representing the capabilities of the FMS API. FMS clients implement this trait.