Struct rusoto_fms::FmsClient

pub struct FmsClient { /* fields omitted */ }

A client for the FMS API.

Methods

impl FmsClient

pub fn new(region: Region) -> FmsClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> FmsClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,