Trait rusoto_fms::Fms

pub trait Fms {
    fn associate_admin_account(
        &self, 
        input: AssociateAdminAccountRequest
    ) -> RusotoFuture<(), AssociateAdminAccountError>;

    fn delete_notification_channel(
        &self
    ) -> RusotoFuture<(), DeleteNotificationChannelError>;

    fn delete_policy(
        &self, 
        input: DeletePolicyRequest
    ) -> RusotoFuture<(), DeletePolicyError>;

    fn disassociate_admin_account(
        &self
    ) -> RusotoFuture<(), DisassociateAdminAccountError>;

    fn get_admin_account(
        &self
    ) -> RusotoFuture<GetAdminAccountResponse, GetAdminAccountError>;

    fn get_compliance_detail(
        &self, 
        input: GetComplianceDetailRequest
    ) -> RusotoFuture<GetComplianceDetailResponse, GetComplianceDetailError>;

    fn get_notification_channel(
        &self
    ) -> RusotoFuture<GetNotificationChannelResponse, GetNotificationChannelError>;

    fn get_policy(
        &self, 
        input: GetPolicyRequest
    ) -> RusotoFuture<GetPolicyResponse, GetPolicyError>;

    fn list_compliance_status(
        &self, 
        input: ListComplianceStatusRequest
    ) -> RusotoFuture<ListComplianceStatusResponse, ListComplianceStatusError>;

    fn list_policies(
        &self, 
        input: ListPoliciesRequest
    ) -> RusotoFuture<ListPoliciesResponse, ListPoliciesError>;

    fn put_notification_channel(
        &self, 
        input: PutNotificationChannelRequest
    ) -> RusotoFuture<(), PutNotificationChannelError>;

    fn put_policy(
        &self, 
        input: PutPolicyRequest
    ) -> RusotoFuture<PutPolicyResponse, PutPolicyError>;
}

Trait representing the capabilities of the FMS API. FMS clients implement this trait.

Required Methods

fn associate_admin_account(
&self,
input: AssociateAdminAccountRequest
) -> RusotoFuture<(), AssociateAdminAccountError>

Sets the AWS Firewall Manager administrator account. AWS Firewall Manager must be associated with a master account in AWS Organizations or associated with a member account that has the appropriate permissions. If the account ID that you submit is not an AWS Organizations master account, AWS Firewall Manager will set the appropriate permissions for the given member account.

The account that you associate with AWS Firewall Manager is called the AWS Firewall manager administrator account.

fn delete_notification_channel(
&self
) -> RusotoFuture<(), DeleteNotificationChannelError>

Deletes an AWS Firewall Manager association with the IAM role and the Amazon Simple Notification Service (SNS) topic that is used to record AWS Firewall Manager SNS logs.

fn delete_policy(
&self,
input: DeletePolicyRequest
) -> RusotoFuture<(), DeletePolicyError>

Permanently deletes an AWS Firewall Manager policy.

fn disassociate_admin_account(
&self
) -> RusotoFuture<(), DisassociateAdminAccountError>

Disassociates the account that has been set as the AWS Firewall Manager administrator account. You will need to submit an AssociateAdminAccount request to set a new account as the AWS Firewall administrator.

fn get_admin_account(
&self
) -> RusotoFuture<GetAdminAccountResponse, GetAdminAccountError>

Returns the AWS Organizations master account that is associated with AWS Firewall Manager as the AWS Firewall Manager administrator.

fn get_compliance_detail(
&self,
input: GetComplianceDetailRequest
) -> RusotoFuture<GetComplianceDetailResponse, GetComplianceDetailError>

Returns detailed compliance information about the specified member account. Details include resources that are in and out of compliance with the specified policy. Resources are considered non-compliant if the specified policy has not been applied to them.

fn get_notification_channel(
&self
) -> RusotoFuture<GetNotificationChannelResponse, GetNotificationChannelError>

Returns information about the Amazon Simple Notification Service (SNS) topic that is used to record AWS Firewall Manager SNS logs.

fn get_policy(
&self,
input: GetPolicyRequest
) -> RusotoFuture<GetPolicyResponse, GetPolicyError>

Returns information about the specified AWS Firewall Manager policy.

fn list_compliance_status(
&self,
input: ListComplianceStatusRequest
) -> RusotoFuture<ListComplianceStatusResponse, ListComplianceStatusError>

Returns an array of PolicyComplianceStatus objects in the response. Use PolicyComplianceStatus to get a summary of which member accounts are protected by the specified policy.

fn list_policies(
&self,
input: ListPoliciesRequest
) -> RusotoFuture<ListPoliciesResponse, ListPoliciesError>

Returns an array of PolicySummary objects in the response.

fn put_notification_channel(
&self,
input: PutNotificationChannelRequest
) -> RusotoFuture<(), PutNotificationChannelError>

Designates the IAM role and Amazon Simple Notification Service (SNS) topic that AWS Firewall Manager uses to record SNS logs.

fn put_policy(
&self,
input: PutPolicyRequest
) -> RusotoFuture<PutPolicyResponse, PutPolicyError>

Creates an AWS Firewall Manager policy.